+++ playwright.config.ts
@@ -3,6 +3,7 @@ import { defineConfig, devices } from "@playwright/test";
 export default defineConfig({
   testDir: "./e2e",
   outputDir: "./e2e/_test-results",
+
   projects: [
     {
       name: "vite-chromium-dist",
@@ -15,7 +16,7 @@ export default defineConfig({
   reporter: [["html", { outputFolder: "e2e/_report", open: "never" }]],
   webServer: {
     command:
-      "pnpm --prefix examples/static-vite build && pnpm --prefix examples/static-vite preview --port 4173 --host",
+      "pnpm --prefix tests/static-vite build && pnpm --prefix tests/static-vite preview --port 4173 --host",
     port: 4173,
     reuseExistingServer: !process.env.CI,
   },